requirements.txt
ultralytics==8.3.*
torch>=2.11.0
torchvision>=0.26.0
matplotlib>=3.10.0
pillow>=12.0.0
modal>=0.73
